requirements.txt
fastapi==0.89.1
httpx==0.23.3
python_bcrypt==0.3.2
requests==2.22.0
uvicorn==0.13.3
